backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
astroid==2.5
attrs==24.3.0
beautifulsoup4==4.12.3
bibtexparser==1.4.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crossref-commons==0.0.7
cryptography==44.0.0
ecdsa==0.19.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
future==1.0.0
geopandas==1.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
HTMLParser==0.0.2
httplib2==0.22.0
idna==2.10
isort==5.13.2
joblib==1.4.2
lazy-object-proxy==1.10.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.6.1
nltk==3.9.1
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20240706
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyChainedProxy==1.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pylint==2.7.1
PyMuPDF==1.25.1
pyogrio==0.10.0
PyPaperBot==1.4.1
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pyproj==3.7.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
ratelimit==2.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
selenium==4.27.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
toml==0.10.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.12.1
wsproto==1.2.0
XlsxWriter==3.2.0
